微调页面，新增登录页logo修改，页脚logo修改

+++ SourseCode/TRM/src/main/java/com/avst/trm/v1/common/conf/Scheduler.java
@@ -227,7 +227,7 @@ public class Scheduler {
                 JexlEngine jexlEngine = new JexlBuilder().create();
                 String formulas=PropertiesListenerConfig.getProperty("record.cache.maxTime");
                 JexlExpression expression = jexlEngine.createExpression(formulas);
-                Integer maxTime = (Integer) expression.evaluate(null);
+                long maxTime = (long) expression.evaluate(null);
 
                 long nowtime=new Date().getTime();
